feat: migrate drizzle to db package
