+++ apps/searchneu/app/api/auth/callback/route.ts
@@ -3,8 +3,7 @@ import { cookies } from "next/headers";
 import { config, googleProvider } from "@/lib/auth/auth";
 import { createJWT } from "@/lib/auth/utils";
 import { decodeIdToken, type OAuth2Tokens } from "arctic";
-import { db } from "@/db";
-import { usersT } from "@/db/schema";
+import { db, usersT } from "@/lib/db";
 import { eq } from "drizzle-orm";
 import { randomUUID } from "node:crypto";
 import { logger } from "@/lib/logger";
