+++ apps/searchneu/app/me/link/page.tsx
@@ -2,8 +2,7 @@ import { config } from "@/lib/auth/auth";
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
 import { verifyJWT } from "@/lib/auth/utils";
-import { db } from "@/db";
-import { usersT } from "@/db/schema";
+import { db, usersT } from "@/lib/db";
 import { eq } from "drizzle-orm";
 import { City } from "@/components/icons/city";
 import { OnboardingFlow } from "@/components/auth/OnboardingFlow";
