+++ apps/searchneu/app/me/trackers/page.tsx
@@ -1,5 +1,4 @@
-import { db } from "@/db";
-import { coursesT, sectionsT, trackersT, usersT } from "@/db/schema";
+import { db, coursesT, sectionsT, trackersT, usersT } from "@/lib/db";
 import { getGuid } from "@/lib/auth/utils";
 import { logger } from "@/lib/logger";
 import { and, eq, isNull } from "drizzle-orm";
