+++ apps/searchneu/lib/auth/onboarding-actions.ts
@@ -3,8 +3,7 @@
 import { cookies } from "next/headers";
 import { config } from "../auth/auth";
 import { verifyJWT } from "../auth/utils";
-import { db } from "@/db";
-import { usersT } from "@/db/schema";
+import { db, usersT } from "@/lib/db";
 import { eq } from "drizzle-orm";
 import { twilio } from "../twilio";
 import { logger } from "../logger";
