+++ apps/searchneu/lib/db.ts
@@ -1,0 +1,4 @@
+import { getDb } from "@sneu/db/client";
+
+export const db = getDb();
+export * from "@sneu/db/schema";
