+++ apps/searchneu/package.json
@@ -15,7 +15,7 @@
     "db:studio": "drizzle-kit studio"
   },
   "dependencies": {
-    "@neondatabase/serverless": "^1.0.2",
+    "@sneu/db": "workspace:*",
     "@opentelemetry/api": "^1.9.0",
     "@tanstack/react-virtual": "^3.13.12",
     "@vercel/analytics": "^1.6.1",
@@ -26,7 +26,6 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
-    "drizzle-orm": "^0.45.0",
     "flags": "^4.0.2",
     "input-otp": "^1.4.2",
     "jose": "^6.1.3",
@@ -50,16 +49,13 @@
     "@types/he": "^1.2.3",
     "@types/mdx": "^2.0.13",
     "@types/node": "^24.10.1",
-    "@types/pg": "^8.15.6",
     "@types/react": "^19.2.7",
     "@types/react-dom": "^19.2.3",
     "babel-plugin-react-compiler": "^1.0.0",
-    "drizzle-kit": "^0.31.8",
     "eslint": "^9.39.1",
     "eslint-config-next": "^16.0.7",
     "he": "^1.2.0",
     "node-html-parser": "^7.0.1",
-    "pg": "^8.16.3",
     "prettier": "^3.7.4",
     "prettier-plugin-tailwindcss": "^0.7.2",
     "server-only": "^0.0.1",
