+++ package.json
@@ -1,9 +1,13 @@
 {
-  "name": "searchmono",
+  "name": "sneu",
   "private": true,
   "scripts": {
     "build": "turbo run build",
-    "dev": "turbo run dev"
+    "dev": "turbo run dev",
+    "db:generate": "turbo run db:generate --filter=@repo/database",
+    "db:migrate": "turbo run db:migrate --filter=@repo/database",
+    "db:push": "turbo run db:push --filter=@repo/database",
+    "db:studio": "turbo run db:studio --filter=@repo/database"
   },
   "devDependencies": {
     "turbo": "^2.6.3"
