+++ packages/db/src/schema/index.ts
@@ -1,0 +1,4 @@
+// unify the schemas into a single import
+export * from "./catalog";
+export * from "./platform";
+export * from "./scheduler";
