+++ packages/tsconfig/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@sneu/tsconfig",
+  "version": "0.0.1",
+  "private": true,
+  "files": [
+    "*.json"
+  ]
+}
